+++ stbt/__init__.py
@@ -16,7 +16,7 @@ from _stbt.core import \
     as_precondition, \
     debug, \
     get_config, \
-    CapturedFrame, \
+    Frame, \
     ConfigurationError, \
     MatchParameters, \
     MatchResult, \
@@ -36,7 +36,7 @@ from _stbt.core import \
 
 __all__ = [
     "as_precondition",
-    "CapturedFrame",
+    "Frame",
     "ConfigurationError",
     "debug",
     "detect_match",
@@ -429,7 +429,7 @@ def frames(timeout_secs=None):
     :returns:
       An ``(image, timestamp)`` tuple for each video frame:
 
-      * ``image`` is a `CapturedFrame` object (that is, an OpenCV image).
+      * ``image`` is a `Frame` (that is, an OpenCV image).
       * ``timestamp`` (deprecated since v26) is a time in nanoseconds. The
         absolute value isn't related to the system time; what's useful is the
         relative difference between frames.
@@ -443,7 +443,7 @@ def frames(timeout_secs=None):
 
 
 def get_frame():
-    """:returns: The latest video frame in OpenCV format (a `CapturedFrame`)."""
+    """:returns: The latest video frame in OpenCV format (a `Frame`)."""
     return _dut.get_frame()